split utils and commands
