+++ e2e/03_registration.spec.js
@@ -4,7 +4,7 @@ import {
   connectToLocalInstance,
   connectToSandbox,
   symfonyConsole,
-} from './utils';
+} from './support/commands';
 
 describe.skip('Registration', () => {
   beforeEach(async () => {
