+++ e2e/foodtech/checkout/customer__role_user__not_logged_in/success.spec.js
@@ -4,7 +4,7 @@ import {
   connectToLocalInstance,
   connectToSandbox,
   symfonyConsole,
-} from '../../../utils';
+} from '../../../support/commands';
 
 describe('checkout for customer with existing account (role - user); not logged in yet; payment - cash on delivery', () => {
   beforeEach(async () => {
